pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vesselme"
version = "0.1.0"
description = "Fundus vessel segmentation annotation tool"
readme = "docs/README.md"
requires-python = ">=3.10"
dependencies = [
  "PySide6>=6.6",
  "numpy>=1.24",
  "opencv-python>=4.8",
  "Pillow>=10.0"
]

[project.scripts]
vesselme = "vesselme.main:run"

[tool.setuptools]
packages = ["vesselme", "vesselme.ui", "vesselme.core", "vesselme.data", "vesselme.services", "vesselme.utils"]
